+++ cmd/influx-stress/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/influxdata/influxdb/stress/cmd"
+
+func main() {
+	cmd.Execute()
+}