+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- [#7611](https://github.com/influxdata/influxdb/pull/7611): Change influx_stress to influx-stress.
 - [#7066](https://github.com/influxdata/influxdb/issues/7066): Add support for secure transmission via collectd.
 - [#7554](https://github.com/influxdata/influxdb/pull/7554): update latest dependencies with Godeps.
 - [#7368](https://github.com/influxdata/influxdb/pull/7368): Introduce syntax for marking a partial response with chunking.